improve python bindings
