+++ python/flight-fusion/Cargo.toml
@@ -8,8 +8,11 @@ edition = "2021"
 keywords = ["datafusion", "datalake", "pandas", "arrow"]
 
 [lib]
-name = "flight_fusion"
+name = "_internal"
 crate-type = ["cdylib"]
+
+[package.metadata.maturin]
+name = "flight_fusion._internal"
 
 [dependencies]
 chrono = "0"
@@ -19,6 +22,7 @@ prost = "0.9"
 # for binary wheel best practice, statically link openssl
 reqwest = { version = "*", features = ["native-tls-vendored"] }
 serde_json = "1"
+thiserror = "1"
 tokio = { version = "1", features = ["rt-multi-thread"] }
 
 # workspace dependencies
