+++ python/flight-fusion/pyproject.toml
@@ -1,20 +1,16 @@
 [project]
 name = "flight-fusion"
-description = "Client to filight fusion serice"
+description = "Client to filight fusion service"
 readme = "README.md"
 license = { file = "LICENSE.txt" }
 requires-python = ">=3.6"
-keywords = ["deltalake", "delta", "datalake", "pandas", "arrow"]
+keywords = ["deltalake", "delta", "arrow"]
 classifiers = [
   "Development Status :: 3 - Alpha",
   "License :: OSI Approved :: Apache Software License",
   "Programming Language :: Python :: 3 :: Only",
 ]
-dependencies = [
-  "pyarrow>=4",
-  'numpy<1.20.0;python_version<="3.6"',
-  'dataclasses;python_version<="3.6"',
-]
+dependencies = ["pyarrow>=4"]
 
 [project.optional-dependencies]
 devel = [
